totalShares is better than tokenizedSharePercentage
Having a variable with the tokenized percentage requires an adjustment and therefore an additional transaction whenever additional shares are tokenized. It is much better to just specify the total number of outstanding shares. The fraction of tokenized shares can be calculated from that. Also, a uint16 is really imprecise. For example, it can't capture 33.3% of the shares being tokenized.